requirements.txt
# requirements.txt

# Core
python-dotenv
loguru

# NLP & Transformers
transformers
sentence-transformers
torch
datasets
scikit-learn
sentencepiece

# Vector DB
faiss-cpu

# Google Firestore
firebase-admin

# ADK (Google Agentic Dev Kit)
google-adk

# Jupyter & Experimentation
jupyterlab
ipykernel
notebook

# Streamlit UI
streamlit

# Others
tqdm
requests
pandas
PyMuPDF
fastapi
uvicorn

#Evaluations
evaluate
bert-score
rouge_score
